requirements.txt
anthropic
textual
python-dotenv